requirements.txt
fastapi
anthropic
scikit-learn
python-dotenv
